WFH-Tasks/wfh_project/requirements.txt
Django==2.2.12
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
pytz==2019.3
sqlparse==0.3.1
